finished refactoring code and renaming filed by exercise number
